+++ apps/api/tests/index.ts
@@ -20,7 +20,7 @@
 import assert from 'node:assert';
 import { after, describe, it } from 'node:test';
 
-import { port } from '../server.js';
+import { port } from '../src/server.js';
 
 describe('Initial Test Setup', () => {
 	describe('First File', () => {
