+++ apps/api/tsconfig.json
@@ -6,5 +6,5 @@
 		"module": "NodeNext",
 		"moduleResolution": "NodeNext"
 	},
-	"include": ["src", "./drizzle.config.ts"]
+	"include": ["src", "./drizzle.config.ts", "tests"]
 }
